refactor(CIV-2488): controller rename
