+++ src/main/routes/routes.ts
@@ -59,7 +59,7 @@ import defendantTimelineController from './features/response/timelineOfEvents/de
 import iDontWantFreeMediationController from './features/mediation/iDontWantFreeMediationController';
 import evidenceController from './features/response/evidence/evidenceController';
 import checkAnswersController from './features/response/checkAnswersController';
-import youHavePaidLessController from './features/response/rejection/fullReject/youHavePaidLess.controller';
+import youHavePaidLessController from './features/response/rejection/fullReject/youHavePaidLessController';
 
 export default [
   homeController,
